feat(logging): Integrate logging system and replace early serial output
- Added a new logging module to handle log messages throughout the kernel.
- Replaced early serial output with logging calls for better consistency and flexibility.
- Updated kernel initialization to use the new logging system, providing detailed boot logs.
- Enhanced memory allocator to ensure proper alignment and block size management.
- Improved panic handling to enter a panic mode for safe logging during panic situations.
- Updated VGA module to use the new logging functions for early boot messages.
- Enabled SSE and SSE2 features in the build configuration for improved performance.

+++ src/interrupts/mod.rs
@@ -87,9 +87,9 @@ fn init_idt() {
         crate::vga::early_serial_str("[IDT] hw int\n");
         
         // ハードウェア割り込みハンドラの設定
-        idt[InterruptVector::Timer as usize].set_handler_fn(timer_interrupt_handler);
-        idt[InterruptVector::Keyboard as usize].set_handler_fn(keyboard_interrupt_handler);
-        idt[InterruptVector::Com1 as usize].set_handler_fn(com1_interrupt_handler);
+        idt[InterruptVector::Timer as u8].set_handler_fn(timer_interrupt_handler);
+        idt[InterruptVector::Keyboard as u8].set_handler_fn(keyboard_interrupt_handler);
+        idt[InterruptVector::Com1 as u8].set_handler_fn(com1_interrupt_handler);
         
         crate::vga::early_serial_str("[IDT] load\n");
